remove exportprocessors list component
